requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3

# Payment Integration
masumi==0.1.10

# File Processing
python-multipart==0.0.18
pypdf2==3.0.1

# AI/ML Libraries
google-generativeai==0.8.3
faiss-cpu==1.9.0
numpy==1.26.4

# HTTP Client
httpx==0.28.1

# Optional but useful for production
gunicorn==23.0.0